+++ packages/react/src/themes/types.ts
@@ -585,6 +585,7 @@ export type ThemeIconSpec = {
 }
 
 export type RequiredIconNames =
+  | 'stardust-checkmark'
   | 'stardust-circle'
   | 'stardust-close'
   | 'stardust-arrow-end'